+++ packages/eslint-plugin/src/configs/strict-type-checked.ts
@@ -5,6 +5,8 @@
 // For developers working in the typescript-eslint monorepo:
 // You can regenerate it using `yarn generate:configs`
 
+import type { ClassicConfig } from '@typescript-eslint/utils/ts-eslint';
+
 export = {
   extends: ['./configs/base', './configs/eslint-recommended'],
   rules: {
@@ -74,4 +76,4 @@ export = {
     '@typescript-eslint/unbound-method': 'error',
     '@typescript-eslint/unified-signatures': 'error',
   },
-};
+} satisfies ClassicConfig.Config;