+++ packages/eslint-plugin/src/rules/no-inferrable-types.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/internal/prefer-ast-types-enum */
 import type { TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
-import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import { createRule } from '../util';
 
@@ -49,8 +48,6 @@ export default createRule<Options, MessageIds>({
     },
   ],
   create(context, [{ ignoreParameters, ignoreProperties }]) {
-    const sourceCode = getSourceCode(context);
-
     function isFunctionCall(
       init: TSESTree.Expression,
       callName: string,
@@ -226,7 +223,7 @@ export default createRule<Options, MessageIds>({
               node.left.optional) ||
             (node.type === AST_NODE_TYPES.PropertyDefinition && node.definite)
           ) {
-            yield fixer.remove(sourceCode.getTokenBefore(typeNode)!);
+            yield fixer.remove(context.sourceCode.getTokenBefore(typeNode)!);
           }
           yield fixer.remove(typeNode);
         },