+++ packages/eslint-plugin/src/rules/triple-slash-reference.ts
@@ -1,6 +1,5 @@
 import type { TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES, AST_TOKEN_TYPES } from '@typescript-eslint/utils';
-import { getSourceCode } from '@typescript-eslint/utils/eslint-utils';
 
 import { createRule } from '../util';
 
@@ -56,7 +55,7 @@ export default createRule<Options, MessageIds>({
   ],
   create(context, [{ lib, path, types }]) {
     let programNode: TSESTree.Node | undefined;
-    const sourceCode = getSourceCode(context);
+
     const references: {
       comment: TSESTree.Comment;
       importName: string;
@@ -97,7 +96,8 @@ export default createRule<Options, MessageIds>({
         programNode = node;
         const referenceRegExp =
           /^\/\s*<reference\s*(types|path|lib)\s*=\s*["|'](.*)["|']/;
-        const commentsBefore = sourceCode.getCommentsBefore(programNode);
+        const commentsBefore =
+          context.sourceCode.getCommentsBefore(programNode);
 
         commentsBefore.forEach(comment => {
           if (comment.type !== AST_TOKEN_TYPES.Line) {